requirements.txt
# Dev environment
pip
autopep8

# App
flask
datasets 
psycopg[binary]
python-dotenv